Vectors and Matrices get an exception in hashing and writing in that they are not considered individual constituents: both represent something too simple and concrete to potentially change meaning down the line, as well as having direct equality comparisons. Building out systems to observe what a save looks like, builds at this point.

+++ Source/Entities/SoundContainer.cpp
@@ -201,8 +201,8 @@ int SoundContainer::Save(Writer& writer) const {
 	return 0;
 }
 
-int SoundContainer::Write(Writer& writer, const Entity& reference, const HashingData& hashData) const {
-	int constituentsConsumed = Entity::Write(writer, reference, hashData);
+size_t SoundContainer::Write(Writer& writer, const Entity& reference, const HashingData& hashData) const {
+	size_t constituentsConsumed = Entity::Write(writer, reference, hashData);
 
 	const SoundContainer& soundSetReference = static_cast<const SoundContainer&>(reference);
